feature(billing): implement Module 4 - Facturacion y Clientes
Update CashReceipt (recibo_caja) to match DB schema: add sede_id,
fp_id, rc_subtotal, rc_descuento, rc_propina, rc_monto_rec, rc_cambio,
auditing fields, CashReceiptStatus enum (ACTIVO/ANULADO), and make
cli_id nullable. Migrate monetary fields from Double to BigDecimal.

Create full CRUD module for CashReceiptDetail (recibo_caja_detalle)
following Clean Architecture: entity, domain model, mapper, repository,
use cases, DTOs, and controller.

+++ src/main/java/development/v/development/application/features/CashReceipt/DeleteCashReceipt/DeleteCashReceiptUseCase.java
@@ -10,16 +10,16 @@ import development.v.development.domain.responses.DataResultDto;
 
 @Service
 public class DeleteCashReceiptUseCase {
-     private final CashReceiptRepository cashReceiptRepository;
+    private final CashReceiptRepository cashReceiptRepository;
 
     public DeleteCashReceiptUseCase(CashReceiptRepository cashReceiptRepository) {
         this.cashReceiptRepository = cashReceiptRepository;
     }
 
-    public DataResultDto<CashReceipt> execute(Integer id) {
+    public DataResultDto<CashReceipt> execute(Long id) {
         CashReceipt foundCashReceipt = cashReceiptRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException(Message.NOT_FOUND));
-                
+
         cashReceiptRepository.delete(id);
         return DataResultDto.success(foundCashReceipt, Message.SUCCESS);
     }
